chore: move public contract class methods to snapi
